All the changes in the app due to the relation changing

+++ src/main/java/org/example/lab2back/service/CategoryService.java
@@ -2,7 +2,9 @@ package org.example.lab2back.service;
 
 import jakarta.persistence.EntityNotFoundException;
 import org.example.lab2back.entity.CategoryEntity;
+import org.example.lab2back.entity.UserEntity;
 import org.example.lab2back.repository.CategoryRepository;
+import org.example.lab2back.repository.UserRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -10,29 +12,38 @@ import java.util.List;
 @Service
 public class CategoryService {
     CategoryRepository categoryRepository;
+    UserRepository userRepository;
 
-    public CategoryService(CategoryRepository categoryRepository) {
+    public CategoryService(CategoryRepository categoryRepository, UserRepository userRepository) {
         this.categoryRepository = categoryRepository;
+        this.userRepository = userRepository;
     }
-
 
     public List<CategoryEntity> getAllCategories() {
         return categoryRepository.findAll();
     }
 
-    public CategoryEntity createCategory(CategoryEntity category) {
+    public CategoryEntity createCategory(CategoryEntity category, Long userId) {
+        UserEntity user = userRepository.findById(userId)
+                .orElseThrow(() -> new EntityNotFoundException("User not found"));
         CategoryEntity newCategory = new CategoryEntity(category.getName());
-        categoryRepository.save(newCategory);
-        return newCategory;
+        newCategory.setUser(user);
+        return categoryRepository.save(newCategory);
     }
 
 
     public void deleteCategory(Long id) {
+        if(!categoryRepository.existsById(id)) throw new EntityNotFoundException("Category not found");
         categoryRepository.deleteById(id);
     }
 
     public CategoryEntity getById(Long id) {
         return categoryRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Category by id: " + id + " not found"));
+    }
+
+    public List<CategoryEntity> getCategoriesByUserId(Long userId) {
+        if (!userRepository.existsById(userId)) throw new EntityNotFoundException("User not found");
+        return categoryRepository.findAllByUserId(userId);
     }
 }
